// search-for-gif/src/app/data/models/gif-backend.ts
export interface GifBackend {
  alt_text: string;
  analytics: {};
  analytics_response_payload: string;
  bitly_gif_url: string;
  bitly_url: string;
  content_url: string;
  embed_url: string;
  id: string;
  images: {};
  import_datetime: string; //"2023-07-14 21:12:34"
  is_sticker: number;
  rating: string;
  slug: string;
  source: string;
  source_post_url: string;
  source_tld: string;
  title: string;
  trending_datetime: string;
  type: string;
  url: string;
  user: {};
  username: String;
}